+++ wire5/src/commonTest/kotlin/mqtt/wire5/control/packet/SubscribeAcknowledgement.kt
@@ -21,12 +21,12 @@ import kotlin.test.assertEquals
 import kotlin.test.fail
 
 class SubscribeAcknowledgementTests {
-    private val packetIdentifier = 2.toUShort()
+    private val packetIdentifier = 2
 
     @Test
     fun packetIdentifier() {
         val payload = GRANTED_QOS_0
-        val puback = SubscribeAcknowledgement(packetIdentifier, payload)
+        val puback = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val data = puback.serialize()
         val pubackResult = ControlPacketV5.from(data) as SubscribeAcknowledgement
         assertEquals(pubackResult.variable.packetIdentifier, packetIdentifier)
@@ -36,7 +36,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun grantedQos1() {
         val payload = GRANTED_QOS_1
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), GRANTED_QOS_1)
     }
@@ -44,7 +44,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun grantedQos2() {
         val payload = GRANTED_QOS_2
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), GRANTED_QOS_2)
     }
@@ -52,7 +52,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun unspecifiedError() {
         val payload = UNSPECIFIED_ERROR
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), UNSPECIFIED_ERROR)
     }
@@ -60,7 +60,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun implementationSpecificError() {
         val payload = IMPLEMENTATION_SPECIFIC_ERROR
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), IMPLEMENTATION_SPECIFIC_ERROR)
     }
@@ -68,7 +68,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun notAuthorized() {
         val payload = NOT_AUTHORIZED
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), NOT_AUTHORIZED)
     }
@@ -76,7 +76,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun topicFilterInvalid() {
         val payload = TOPIC_FILTER_INVALID
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), TOPIC_FILTER_INVALID)
     }
@@ -84,7 +84,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun packetIdentifierInUse() {
         val payload = PACKET_IDENTIFIER_IN_USE
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), PACKET_IDENTIFIER_IN_USE)
     }
@@ -92,7 +92,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun quotaExceeded() {
         val payload = QUOTA_EXCEEDED
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), QUOTA_EXCEEDED)
     }
@@ -100,7 +100,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun sharedSubscriptionsNotSupported() {
         val payload = SHARED_SUBSCRIPTIONS_NOT_SUPPORTED
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), SHARED_SUBSCRIPTIONS_NOT_SUPPORTED)
     }
@@ -108,7 +108,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun subscriptionIdentifiersNotSupported() {
         val payload = SUBSCRIPTION_IDENTIFIERS_NOT_SUPPORTED
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), SUBSCRIPTION_IDENTIFIERS_NOT_SUPPORTED)
     }
@@ -116,7 +116,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun wildcardSubscriptionsNotSupported() {
         val payload = WILDCARD_SUBSCRIPTIONS_NOT_SUPPORTED
-        val obj = SubscribeAcknowledgement(packetIdentifier, payload)
+        val obj = SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
         val result = ControlPacketV5.from(obj.serialize()) as SubscribeAcknowledgement
         assertEquals(result.payload.first(), WILDCARD_SUBSCRIPTIONS_NOT_SUPPORTED)
     }
@@ -125,7 +125,7 @@ class SubscribeAcknowledgementTests {
     fun invalidVariableHeaderPayload() {
         val payload = MALFORMED_PACKET
         try {
-            SubscribeAcknowledgement(packetIdentifier, payload)
+            SubscribeAcknowledgement(packetIdentifier.toUShort(), payload)
             fail()
         } catch (e: ProtocolError) {
         }
@@ -134,7 +134,7 @@ class SubscribeAcknowledgementTests {
     @Test
     fun reasonString() {
         val props = VariableHeader.Properties(reasonString = MqttUtf8String("yolo"))
-        val actual = SubscribeAcknowledgement(packetIdentifier, props, GRANTED_QOS_1)
+        val actual = SubscribeAcknowledgement(packetIdentifier.toUShort(), props, GRANTED_QOS_1)
         val bytes = actual.serialize()
         val expected = ControlPacketV5.from(bytes) as SubscribeAcknowledgement
         assertEquals(expected.variable.properties.reasonString, MqttUtf8String("yolo"))
@@ -170,7 +170,7 @@ class SubscribeAcknowledgementTests {
         }
         assertEquals(userPropertyResult.size, 1)
 
-        val request = SubscribeAcknowledgement(packetIdentifier, props, WILDCARD_SUBSCRIPTIONS_NOT_SUPPORTED).serialize()
+        val request = SubscribeAcknowledgement(packetIdentifier.toUShort(), props, WILDCARD_SUBSCRIPTIONS_NOT_SUPPORTED).serialize()
         val requestRead = ControlPacketV5.from(request.copy()) as SubscribeAcknowledgement
         val (key, value) = requestRead.variable.properties.userProperty.first()
         assertEquals(key.getValueOrThrow(), "key")
